Fix a bug in UITelemetry reporting.
In particular, we should only remove the listener if we’re not shutting
down, so that it’s still around when the Telemetry system tries to
gather its data, which is long after we’re disabled.  I don’t know if
we need to remove the function on startup, but I figured better safe
than sorry.  ;)